Big stuff

+++ app/src/main/java/jackwtat/simplembta/MbtaData/Stop.java
@@ -24,7 +24,7 @@ public class Stop implements Comparable {
         this.longitude = 0.0;
     }
 
-    public Stop(String id, String name, float latitude, float longitude) {
+    public Stop(String id, String name, double latitude, double longitude) {
         this.id = id;
         this.name = name;
         this.latitude = latitude;
